Rewrite 'platform.ts' in TypeScript
