+++ package.json
@@ -155,6 +155,7 @@
     "@types/jest": "^26.0.9",
     "@types/lunr": "^2.3.3",
     "@types/node": "^14.0.26",
+    "@types/platform": "^1.3.2",
     "@types/prop-types": "^15.7.3",
     "@types/react": "^16.9.46",
     "@types/react-bootstrap": "^0.32.22",
